made changes in category controller
