+++ src/main/java/com/cg/fds/entities/Item.java
@@ -17,7 +17,7 @@ public class Item {
 	private String itemId;
 	private String itemName;
 	
-	@OneToOne
+	@ManyToOne
 	private Category category;
 	
 	private int quantity;
