+++ src/main/java/com/cg/fds/service/CategoryServiceImp.java
@@ -2,6 +2,7 @@ package com.cg.fds.service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.Random;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,14 +20,27 @@ import com.cg.fds.repository.IItemRepository;
 
 @Service
 public class CategoryServiceImp implements ICategoryService {
+	
 	@Autowired
 	private IItemRepository itemRepository;
 	@Autowired
 	private ICategoryRepository categoryRepository;
+	
+	public String generateId() {
+		StringBuilder builder= new StringBuilder();
+		Random random= new Random();
+		for(int i=0; i<10; i++) {
+			int randomNum=random.nextInt(10);
+			builder.append(randomNum);
+		}
+		return builder.toString();
+	}
 
 	@Override
 	public Category addCategory(Category cat) {
 		validateCategory(cat);
+		String id = generateId();
+		cat.setCatId(id);
 		return categoryRepository.save(cat);
 	}
 
@@ -40,20 +54,20 @@ public class CategoryServiceImp implements ICategoryService {
 		return categoryRepository.save(cat);
 	}
 
-	@Override
-	public Category removeCategory(Category cat) {
-		validateCategory(cat);
-		boolean exists = categoryRepository.existsById(cat.getCatId());
-		if (!exists) {
-			throw new RemoveCategoryException("Category with id not present=" + cat.getCatId());
-		}
-		List<Item> items = itemRepository.findByCategory(cat);
-		for(Item item:items) {
-			itemRepository.delete(item);
-		}
-		categoryRepository.delete(cat);
-		return cat;
-	}
+//	@Override
+//	public Category removeCategory(Category cat) {
+//		validateCategory(cat);
+//		boolean exists = categoryRepository.existsById(cat.getCatId());
+//		if (!exists) {
+//			throw new RemoveCategoryException("Category with id not present=" + cat.getCatId());
+//		}
+//		List<Item> items = itemRepository.findByCategory(cat);
+//		for(Item item:items) {
+//			itemRepository.delete(item);
+//		}
+//		categoryRepository.delete(cat);
+//		return cat;
+//	}
 
 	@Override
 	public Category viewCategory(String catId) {
