+++ src/main/java/com/cg/fds/service/ICategoryService.java
@@ -8,7 +8,7 @@ public interface ICategoryService {
 
 	public Category addCategory(Category cat);
 	public Category updateCategory(Category cat);
-	public Category removeCategory(Category cat);
+//	public Category removeCategory(Category cat);
 	public Category viewCategory(String cat);
 	public List<Category> viewAllCategory();
 }
