+++ src/main/java/com/cg/fds/util/CategoryUtil.java
@@ -15,15 +15,7 @@ public class CategoryUtil {
 	public Category getCategory() {
 		return new Category();
 	}
-	public String generateId() {
-		StringBuilder builder= new StringBuilder();
-		Random random= new Random();
-		for(int i=0; i<10; i++) {
-			int randomNum=random.nextInt(10);
-			builder.append(randomNum);
-		}
-		return builder.toString();
-	}
+	
 public CategoryDetails toCategoryDetail(Category category) {
 	CategoryDetails cd = new CategoryDetails();
 	cd.setName(category.getCategoryName());
